+++ plugins/rplanners/pathoptimizers.cpp
@@ -30,6 +30,7 @@ public:
         EnvironmentMutex::scoped_lock lock(GetEnv()->GetMutex());
         _parameters.reset(new TrajectoryTimingParameters());
         _parameters->copy(params);
+        _probot = pbase;
         return _InitPlan();
     }
 
@@ -38,6 +39,7 @@ public:
         EnvironmentMutex::scoped_lock lock(GetEnv()->GetMutex());
         _parameters.reset(new TrajectoryTimingParameters());
         isParameters >> *_parameters;
+        _probot = pbase;
         return _InitPlan();
     }
 
@@ -62,6 +64,12 @@ public:
         if( ptraj->GetNumWaypoints() < 2 ) {
             return PS_Failed;
         }
+
+        RobotBase::RobotStateSaverPtr statesaver;
+        if( !!_probot ) {
+            statesaver.reset(new RobotBase::RobotStateSaver(_probot));
+        }
+
         uint32_t basetime = GetMilliTime();
         PlannerParametersConstPtr parameters = GetParameters();
 
@@ -165,6 +173,7 @@ protected:
     }
 
     TrajectoryTimingParametersPtr _parameters;
+    RobotBasePtr _probot;
 };
 
 //    virtual void _OptimizePathSingle(list<Node*>& path, int numiterations)